Back to using `regprocedure` for routine signature
See https://dba.stackexchange.com/a/324899/79909 for why this a safe
thing to do.

+++ bin/test_dump_restore.sh
@@ -6,8 +6,11 @@ PG_BIN_DIR="$(pg_config --bindir)"
 usage() {
     cat <<EOF
 Usage:
-    $SCRIPT_NAME --extension <extension_name> --psql-script-file <file> --out-file <file> --expected-out-file <file>
+    $SCRIPT_NAME [options] --extension <extension_name> --psql-script-file <file> --out-file <file> --expected-out-file <file>
     $SCRIPT_NAME --help|-h
+
+Options:
+    --keep-temp-dir
 EOF
 }
 
@@ -15,6 +18,7 @@ psql_script_file=""
 expected_out_file=""
 out_file=""
 extension_name=""
+keep_temp_dir=""
 while [[ "$#" -gt 0 ]]; do
     case "$1" in
         -h|--help)
@@ -36,6 +40,10 @@ while [[ "$#" -gt 0 ]]; do
         --extension)
             extension_name="$2"
             shift 2
+            ;;
+        --keep-temp-dir)
+            keep_temp_dir="alright"
+            shift
             ;;
         --*|-*)
             echo -e "\e[31mUnrecognized option: \e[1m$1\e[0m" >&2
@@ -71,7 +79,11 @@ cleanup() {
     fi
 
     if [ -n "$tmp_dir" ]; then
-        rm -rf "$tmp_dir"
+        if [ -n "$keep_temp_dir" ]; then
+            echo -e "Temp. dir. preserved: \e[1m$tmp_dir\e[22m"
+        else
+            rm -rf "$tmp_dir"
+        fi
     fi
 }
 trap cleanup exit
@@ -85,6 +97,7 @@ export PGHOST="$tmp_dir"
 export PGDATABASE="test_dump_restore"
 export PGUSER="wortel"
 export PGVERSION="15"
+OID_NOISE_DB_NAME="test_dump_restore_oid_noise"
 
 start_pg_cluster() {
     "$PG_BIN_DIR/initdb" \
@@ -136,8 +149,19 @@ $PG_BIN_DIR/psql postgres -c '\set ON_ERROR_STOP' \
     >> "$out_file" 2>&1 \
     || exit 5
 
+echo "-- createdb '$OID_NOISE_DB_NAME'" >> "$out_file"
+$PG_BIN_DIR/createdb "$OID_NOISE_DB_NAME" >> "$out_file" 2>&1 || exit 5
+
+echo "-- psql -f '$psql_script_file' -v 'extension_name=$extension_name' -v 'test_stage=pre-restore'" >> "$out_file"
+$PG_BIN_DIR/psql \
+    -f "$psql_script_file" \
+    -v "extension_name=$extension_name" \
+    -v "test_stage=pre-restore" \
+    "$OID_NOISE_DB_NAME" \
+    >> "$out_file" 2>&1 || exit 5
+
 echo "-- pg_restore --create --dbname postgres <dump_file>" >> "$out_file"
-$PG_BIN_DIR/pg_restore --create --dbname postgres "$dump_file" >> "$out_file" 2>&1 || exit 5
+$PG_BIN_DIR/pg_restore --exit-on-error --create --dbname postgres "$dump_file" >> "$out_file" 2>&1 || exit 5
 
 echo "-- psql -f '$psql_script_file' -v 'extension_name=$extension_name' -v 'test_stage=post-restore'" >> "$out_file"
 $PG_BIN_DIR/psql \
